requirements.txt
Flask
Fabric
tinydb
Flask-SQLAlchemy
Flask-WTF
WTForms
coverage
asyncio
ccxt
tinydb
python-dateutil
pandas
requests
python-dotenv